\name{Availabledata}
\alias{Availabledata}
\docType{data}
\title{
    Available datasources
}
\description{
    Availabledata contains a character vector with the names of the
    datasources that the package contains.
}
\usage{data("Availabledata")}
\format{
    A character vector with the names of the datasources.
}
\examples{
    data(Availabledata)
    cat(Availabledata)
}
\keyword{datasets}
